TokamakNeutronSource

Documentation for TokamakNeutronSource.

TokamakNeutronSource.TokamakNeutronSource

TokamakNeutronSource.TokamakNeutronSource — Module

TokamakNeutronSource

Code to compute neutron emission intensity for given plasma state.

Description

The plasma state is presented with:

ψ(r,z) - magnetic surface coordinate spatial distribution

n(ψ) - ion concentration in dependency of magnetic surface coordinate ψ

T(ψ) - Tritium concentration in dependency of magnetic surface coordinate ψ

r - ratio of atomic concentraions of tritium to deuterium one